functionnal dashboard prototype
