+++ package.json
@@ -22,6 +22,7 @@
     "react": "^18.2.0",
     "react-apexcharts": "^1.3.7",
     "react-bootstrap": "^2.10.2",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-router": "^6.3.0",
